requirements.txt
Django==3.1.6
djangorestframework==3.14.0
mysqlclient==2.2.0
celery==5.3.4
